feat(ds): add vector data structure
With this commit we add a first implementation of a vector,
following as closed as possible the interface described
in cppreference. Nonetheless, please keep in mind that we
had made no optimizations yet; in particular, we have not used
any kind of `constexpr` definition for any method. That also
means that we have followed the C++17 definitions and not the
C++20.

The commit also includes the respective tests. There is also
room for work there, but at least we tried to test all methods at
least once.